Main<-Exempel

+++ Hattmakare/Models/Hats/MaterialQuantityViewModel.cs
@@ -6,6 +6,6 @@
         public string Name { get; set; }
         public string Unit { get; set; }
         public decimal Price { get; set; }
-        public int Quantity { get; set; } = 0;
+        public double Quantity { get; set; } = 0;
     }
 }
